+++ packages/client-auth/src/lib/config.ts
@@ -59,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.22",
-    genVersion: "2.387.0",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.22 2.387.0 0.0.0 @formance/sdk-auth",
+    sdkVersion: "1.0.0-alpha.23",
+    genVersion: "2.390.6",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.23 2.390.6 0.0.0 @formance/sdk-auth",
 } as const;